barter-dex-api: using new api
