+++ app/containers/BuyPage/saga/load-buy-coin-process.js
@@ -2,7 +2,7 @@ import swal from 'sweetalert';
 // import { put, call, select, cancel, cancelled } from 'redux-saga/effects';
 import { put, call, select, cancelled } from 'redux-saga/effects';
 import { delay } from 'redux-saga';
-import api from '../../../utils/barter-dex-api/index2';
+import api from '../../../utils/barter-dex-api/index';
 import {
   makeSelectCurrentUser,
   makeSelectBalanceEntities
