+++ app/utils/barter-dex-api/is-ready.saved
@@ -1,0 +1,18 @@
+isReady() {
+  return new Promise((resolve, reject) => {
+    const interval = setInterval(async () => {
+      try {
+        debug('isReady');
+        await this.fetch.find({});
+        clearInterval(interval);
+        resolve('OK');
+        // eslint-disable-next-line no-empty
+      } catch (_) {}
+    }, 100);
+
+    setTimeout(() => {
+      clearInterval(interval);
+      reject(new Error('Giving up trying to connect to marketmaker'));
+    }, 10000);
+  });
+}
