--- app/utils/barter-dex-api/pendingswaps.js
@@ -1,3 +1,0 @@
-// @flow
-
-// https://docs.komodoplatform.com/barterDEX/barterDEX-API.html?highlight=listunspent#pendingswaps
